Fix/cart price data type

+++ carts/schemas.py
@@ -1,10 +1,11 @@
+from decimal import Decimal
 from typing import List, Optional
 from ninja import Schema, Field
 
 
 class CartIn(Schema):
     book_id: int
-    price: Optional[int]
+    price: Optional[Decimal]
     amount: int
 
 
@@ -16,13 +17,13 @@ class Book(Schema):
 class CartItem(Schema):
     id: int
     book: Book
-    price: int = Field(..., alias='unit_price')
+    price: Decimal = Field(..., alias='unit_price')
     amount: int
 
 
 class CartOut(Schema):
     data: List[CartItem]
-    total_price: float
+    total_price: Decimal
 
 
 class DeleteCartIn(Schema):
@@ -35,11 +36,11 @@ class CheckoutCartIn(Schema):
 
 class OrderedItem(Schema):
     book: Book
-    price: int = Field(..., alias='unit_price')
+    price: Decimal = Field(..., alias='unit_price')
     amount: int
 
 
 class CheckoutCartOut(Schema):
     order_id: int = Field(..., alias='id')
-    total_price: int
+    total_price: Decimal
     items: List[OrderedItem] = Field(..., alias='ordereditem_set')
